Update activity to fragment

+++ app/src/main/java/vn/mran/bc3/model/RuleMain.java
@@ -16,7 +16,7 @@ public class RuleMain {
         this.status = status;
     }
 
-    public int getQuantum() {
-        return (int) quantum;
+    public long getQuantum() {
+        return  quantum;
     }
 }
